% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{compute_features_4}
\alias{compute_features_4}
\title{Computes features for each network}
\usage{
compute_features_4(gr, attributes = FALSE, attr_name = NULL)
}
\description{
Computes features for each network
}
