% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laplacian_eigenvalue_method.R
\name{lad}
\alias{lad}
\title{Laplacian Eigen Value method by Shenyang Huang, Yasmeen Hitti, Guillaume Rabusseau
and Reihaneh Rabbany from their KDD'20 paper Laplacian Change Point Detection for Dynamic Graphs}
\usage{
lad(matlist, k = NULL, short_win, long_win, alpha = 0.05, from_file = NULL)
}
\arguments{
\item{matlist}{The matrix list, where each matrix is an adjacency matrix of the graph.}

\item{k}{The number of eigen values to connsider}

\item{short_win}{The length of the shorter windows}

\item{long_win}{The length of the longer windows}

\item{alpha}{The threshold to declare anomalies}

\item{from_file}{This is an additional parameter only if a file needs to be read}
}
\description{
Laplacian Eigen Value method by Shenyang Huang, Yasmeen Hitti, Guillaume Rabusseau
and Reihaneh Rabbany from their KDD'20 paper Laplacian Change Point Detection for Dynamic Graphs
}
\examples{
library(VCERGM)
networks <- Rollcall$networks
ladobj <- lad(networks, short_win = 5, long_win = 10)

}
\references{
Huang, S., Hitti, Y., Rabusseau, G., & Rabbany, R. (2020). Laplacian Change
Point Detection for Dynamic Graphs. Proceedings of the ACM SIGKDD International Conference
on Knowledge Discovery and Data Mining, 349–358. https://doi.org/10.1145/3394486.3403077
}
