% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oddnet.R
\docType{package}
\name{oddnet-package}
\alias{oddnet-package}
\title{oddnet: Anomaly Detection in Temporal Networks}
\description{
Anomaly detection in dynamic, temporal networks. Use four spaces when indenting paragraphs within the Description.
}
\author{
\strong{Maintainer}: Sevvandi Kandanaarachchi \email{sevvandik@gmail.com} (\href{https://orcid.org/0000-0002-0337-0395}{ORCID})

Authors:
\itemize{
  \item Rob Hyndman \email{rob.hyndman@monash.edu} (\href{https://orcid.org/0000-0002-2140-5352}{ORCID})
}

}
\keyword{internal}
