% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensorsplat.R
\name{tensorsplat}
\alias{tensorsplat}
\title{Implements Danai Koutra's TensorSplat algorithm}
\usage{
tensorsplat(matlist, k = 2, alpha = 0.05)
}
\arguments{
\item{matlist}{The list of matrices where each matrix denotes the adjacency matrix of the network.}

\item{k}{The number of components in PARFAC tensor decomposition.}

\item{alpha}{The threshold to declare anoalies}
}
\description{
Implements Danai Koutra's TensorSplat algorithm
}
\references{
Koutra, D., Papalexakis, E. E., & Faloutsos, C. (2012).
TensorSplat: Spotting latent anomalies in time. Proceedings of the 2012 16th
Panhellenic Conference on Informatics, PCI 2012, 144–149.
https://doi.org/10.1109/PCi.2012.60
}
