% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oddnet.R
\docType{package}
\name{oddnet-package}
\alias{oddnet-package}
\title{oddnet: Anomaly Detection in Temporal Networks}
\description{
Anomaly detection in dynamic, temporal networks. The package 'oddnet' uses a feature-based method to identify anomalies. First, it computes many features for each network. Then it models the features using time series methods. Using time series residuals it detects anomalies. This way, the temporal dependencies are accounted for when identifying anomalies (Kandanaarachchi, Hyndman 2022) \href{https://arxiv.org/abs/2210.07407}{arXiv:2210.07407}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://sevvandi.github.io/oddnet/}
}

}
\author{
\strong{Maintainer}: Sevvandi Kandanaarachchi \email{sevvandik@gmail.com} (\href{https://orcid.org/0000-0002-0337-0395}{ORCID})

Authors:
\itemize{
  \item Rob Hyndman \email{rob.hyndman@monash.edu} (\href{https://orcid.org/0000-0002-2140-5352}{ORCID})
}

}
\keyword{internal}
